backend/requirements.txt
django
djangorestframework
pandas
boto3
duckdb
django-cors-headers
gunicorn